# openapi.yaml
openapi: 3.0.3
info:
  title: 转型罗盘 TranSiCompass API
  version: 1.0.0
  description: |
    工程人转型决策工具的后端契约。前端据此生成 TypeScript 类型与 Mock，后端据此实现。

    约定：
    - 业务错误一律返回 HTTP 200 + 非 0 `code`（规避微信内置浏览器与企业网络对 4xx/5xx 的拦截改写）。
      仅网关级失败（限流、超大 body）返回真实 4xx/5xx。
    - 服务端不存在任何接收 multipart/form-data 的端点。简历解析与 PII 脱敏在客户端完成，
      仅上送 sanitized_text。这是合规硬约束，见 Spec 8.5。
    - 所有诊断数据默认 30 天自动清除；分享链接 7 天有效。

  contact:
    name: 架构负责人 高见远
  license:
    name: Proprietary
servers:
  - url: https://{host}/api/v1
    description: 生产
    variables:
      host:
        default: transicompass.example.cn
  - url: http://localhost:3000/api/v1
    description: 本地开发
...
security:
  - sessionCookie: []
paths:
...
  /results/{id}:
    get:
      tags: [result]
      operationId: getResult
      summary: 读取结果（本人 / 已登录 / 分享令牌 三选一）
      security:
        - sessionCookie: []
        - jwtCookie: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: share_token
          in: query
          required: false
          description: 只读分享场景传入；命中时 view_mode 为 shared，服务端强制剥离敏感字段
          schema: { type: string, minLength: 32, maxLength: 32 }
      responses:
        '200':
          description: 统一响应
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ResultView'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: 匿名会话 Cookie（httpOnly, Secure, SameSite=Lax, 30d），不含实名信息
    jwtCookie:
      type: apiKey
      in: cookie
      name: token
      description: 轻登录 JWT（HS256, httpOnly, Secure, SameSite=Lax, 7d）
...
  schemas:
    ApiEnvelope:
      type: object
      required: [code, data, message]
      properties:
        code:
          type: integer
          description: |
            0 成功
            1001 参数校验失败 / 1002 未勾选隐私授权 / 1003 内容过短或过长 / 1004 与工程行业无关
            2001 未登录 / 2002 验证码错误或过期 / 2003 无权访问
            3001 简历解析失败 / 3002 文件过大或格式不支持
            4001 AI 服务不可用（可降级）/ 4002 AI 输出校验失败 / 4003 检测到提示词注入 / 4004 AI 超时
            5001 分享链接不存在或已过期 / 5002 数据已删除
            6001 限流 / 9000 服务器内部错误
          enum: [0, 1001, 1002, 1003, 1004, 2001, 2002, 2003, 3001, 3002,
                 4001, 4002, 4003, 4004, 5001, 5002, 6001, 9000]
        data:
          nullable: true
        message:
          type: string
          description: 面向用户的中文可读描述，不含技术堆栈信息
    TrackCode:
      type: string
      description: V1 锁定 3 个赛道，模型不得生成枚举外的值
      enum:
        - NEW_ENERGY_STORAGE
        - SMART_CONSTRUCTION_BIM
        - ENGINEERING_B2B_OVERSEAS
...
    Evidence:
      type: object
      description: 可解释性的最小单元。source_snippet 必须逐字出自 sanitized_text，不得改写编造。
      required: [source_snippet, why]
      properties:
        source_snippet: { type: string, minLength: 4, maxLength: 120 }
        why: { type: string, minLength: 8, maxLength: 200 }
    Reason:
      type: object
      description: 赛道匹配的结构化依据三元组，是 PIPL 第 24 条说明权与「推理依据占卡片主体」版式的落地点。
      required: [source_quote, mapped_capability, target_scenario]
      properties:
        source_quote:
          type: string
          minLength: 4
          maxLength: 120
          description: 逐字出自 sanitized_text 的原句片段（卡片主体）
        mapped_capability:
          type: string
          minLength: 4
          maxLength: 60
          description: 该佐证映射到的能力 / 经验点
        target_scenario:
          type: string
          minLength: 8
          maxLength: 120
          description: 在目标赛道中的对应场景
    TransferableSkill:
      type: object
      required: [name, description, strength, source_quote, evidence]
      properties:
        name: { type: string, minLength: 2, maxLength: 24 }
        description: { type: string, minLength: 10, maxLength: 160 }
        strength: { type: string, enum: [high, medium, low] }
        source_quote:
          type: string
          minLength: 4
          maxLength: 120
          description: 逐字出自 sanitized_text 的代表性佐证，作为卡片主体（PIPL 24 可解释性落点）
        evidence:
          type: array
          minItems: 1
          maxItems: 3
          items: { $ref: '#/components/schemas/Evidence' }
    TrackMatch:
      type: object
      description: reasons 最少 2 条是硬约束——结构上不允许出现「裸奔的百分比」
      required: [track_code, match_score, match_level, reasons, typical_roles, caveat]
      properties:
        track_code: { $ref: '#/components/schemas/TrackCode' }
        match_score: { type: integer, minimum: 0, maximum: 100 }
        match_level: { type: string, enum: [high, medium, low] }
        reasons:
          type: array
          minItems: 2
          maxItems: 4
          items: { $ref: '#/components/schemas/Reason' }
        typical_roles:
          type: array
          minItems: 3
          maxItems: 6
          items: { type: string, minLength: 2, maxLength: 20 }
        caveat:
          type: string
          minLength: 10
          maxLength: 200
          description: 该方向的风险提醒，用于对冲「伪风口」，是可信度来源
    Gap:
      type: object
      required: [gap_name, why_it_matters, closing_action]
      properties:
        gap_name: { type: string, minLength: 2, maxLength: 24 }
        why_it_matters: { type: string, minLength: 10, maxLength: 200 }
        closing_action:
          type: string
          minLength: 10
          maxLength: 200
          description: 必须是可执行动作，不得写成「多学习」
    PathStage:
      type: object
      required: [stage, deliverable, why_this_deliverable, verifiable_artifact]
      properties:
        stage: { type: string, enum: ['0-1m', '1-3m', '3-6m'] }
        deliverable:
          type: string
          minLength: 10
          maxLength: 120
          description: 必须是能写进简历的具体产出物，禁止写成「学习某某课程」
          example: 独立完成一份 20MW 工商业储能 EPC 项目投标技术方案
        why_this_deliverable: { type: string, minLength: 10, maxLength: 200 }
        verifiable_artifact: { type: string, minLength: 4, maxLength: 80 }
    RewriteSample:
      type: object
      required: [target_track_code, original, rewritten, what_changed]
      properties:
        target_track_code: { $ref: '#/components/schemas/TrackCode' }
        original: { type: string, minLength: 6, maxLength: 200 }
        rewritten: { type: string, minLength: 6, maxLength: 200 }
        what_changed: { type: string, minLength: 6, maxLength: 160 }
    DiagnosisPayload:
      type: object
      required:
        - out_of_scope
        - summary
        - transferable_skills
        - track_matches
        - top_gaps
        - learning_path
        - rewrite_samples
      properties:
        out_of_scope:
          type: boolean
          description: 与三个赛道均无实质关联时为 true，必须给出 reason，禁止编造第四个赛道
        out_of_scope_reason: { type: string, maxLength: 300, nullable: true }
        summary: { type: string, minLength: 20, maxLength: 200 }
        transferable_skills:
          type: array
          minItems: 5
          maxItems: 8
          items: { $ref: '#/components/schemas/TransferableSkill' }
        track_matches:
          type: array
          minItems: 3
          maxItems: 3
          description: 恰好 3 条，track_code 互不相同
          items: { $ref: '#/components/schemas/TrackMatch' }
        top_gaps:
          type: array
          minItems: 2
          maxItems: 2
          items: { $ref: '#/components/schemas/Gap' }
        learning_path:
          type: array
          minItems: 3
          maxItems: 3
          description: 顺序固定为 0-1m, 1-3m, 3-6m
          items: { $ref: '#/components/schemas/PathStage' }
        rewrite_samples:
          type: array
          minItems: 3
          maxItems: 3
          items: { $ref: '#/components/schemas/RewriteSample' }
    ResultView:
      type: object
      nullable: true
      required: [result_id, created_at, status, view_mode, payload, model_disclosure, disclaimer]
      properties:
        result_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        status: { type: string, enum: [completed, degraded] }
        out_of_scope: { type: boolean }
        view_mode:
          type: string
          enum: [owner, shared]
          description: shared 时服务端强制剥离 sanitize_report / session_id / model_meta.raw_usage
        payload: { $ref: '#/components/schemas/DiagnosisPayload' }
        model_disclosure:
          type: object
          description: 合规公示，页面必须渲染（生成式人工智能服务登记要求）
          required: [model_name, registration_no]
          properties:
            model_name: { type: string, example: qwen3.7-plus }
            registration_no: { type: string }
        disclaimer:
          type: string
          example: 本报告为 AI 辅助参考，不构成职业中介服务、就业承诺或投资建议。
...